PDF file uploaded
This is the final report that I had to make for the system. This report contains the schematic that I used to create the project, and then I have 2 FSM diagrams that I hand-written along with the code in asm that I created in order to implement the stated functionality given in the fsm diagrams.